refactor: MainContent에서 FeedPage 분리
