+++ src/pages/MainContent.jsx
@@ -1,8 +1,6 @@
 "use client";
 
-import { useState } from "react";
-import FilterBar from "@/components/Feed/FilterBar";
-import FeedGrid from "@/components/Feed/FeedGrid";
+import FeedPage from "@/pages/FeedPage";
 import PhotoCompanionPage from "@/pages/PhotoCompanionPage";
 import LocalCompanionPage from "@/pages/LocalCompanionPage";
 import ProfileManagementPage from "@/pages/ProfileManagementPage";
@@ -74,36 +72,6 @@ const MainContent = ({
     >
       {renderPage()}
     </main>
-  );
-};
-
-const FeedPage = ({
-  onFeedCountChange,
-  isLoggedIn,
-  feedData,
-  onFeedDelete,
-}) => {
-  const [filters, setFilters] = useState({
-    author: "",
-    title: "",
-    region: "",
-  });
-
-  const handleFilterChange = (newFilters) => {
-    setFilters(newFilters);
-  };
-
-  return (
-    <>
-      <FilterBar filters={filters} onFilterChange={handleFilterChange} />
-      <FeedGrid
-        filters={filters}
-        onFeedCountChange={onFeedCountChange}
-        isLoggedIn={isLoggedIn}
-        feedData={feedData}
-        onFeedDelete={onFeedDelete}
-      />
-    </>
   );
 };
